baidu-index-hunter-backend/requirements.txt
# Web框架
Flask==2.0.1
Flask-Cors==3.0.10
flasgger==0.9.5
flask-socketio==5.1.1

# 数据库
pymysql==1.0.2

# 缓存
redis==3.5.3

# HTTP客户端
requests==2.26.0

# 爬虫相关
selenium>=4.10.0
webdriver-manager>=3.8.0

# Scrapy 框架
Scrapy>=2.11.0
Twisted>=23.10.0
# Scrapy 相关依赖
itemadapter>=0.8.0
parsel>=1.8.0
protego>=0.2.0
w3lib>=2.1.0

# 任务调度
apscheduler>=3.9.0
schedule==1.2.1

# 数据验证 (兼容现有环境)
python-dotenv==0.19.1

# 日志
loguru>=0.6.0

# 其他工具
python-multipart>=0.0.5
aiofiles>=0.8.0

# 其他依赖
Werkzeug==2.0.1
marshmallow==3.13.0
PyJWT==2.1.0
fake-useragent==1.1.1
pandas==1.3.3